webos part-1/requirements.txt
Flask>=2.3.0
Flask-SocketIO>=5.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.1.0
python-dotenv>=1.0.0
PyJWT[crypto]>=2.8.0
bcrypt>=4.0.0
python-socketio>=5.8.0
eventlet>=0.33.3
Werkzeug>=2.3.0 